+++ LDSContent/ItemCategory.swift
@@ -24,10 +24,10 @@ import Foundation
 
 public struct ItemCategory: Equatable {
     
-    public var id: Int
+    public var id: Int64
     public var name: String
     
-    public init(id: Int, name: String) {
+    public init(id: Int64, name: String) {
         self.id = id
         self.name = name
     }
