+++ LDSContent/NavItem.swift
@@ -24,17 +24,17 @@ import Foundation
 
 public struct NavItem: NavNode, Equatable {
     
-    public var id: Int
-    public var navSectionID: Int
+    public var id: Int64
+    public var navSectionID: Int64
     public var position: Int
     public var imageRenditions: [ImageRendition]
     public var titleHTML: String
     public var subtitle: String?
     public var preview: String?
     public var uri: String
-    public var subitemID: Int
+    public var subitemID: Int64
     
-    public init(id: Int, navSectionID: Int, position: Int, imageRenditions: [ImageRendition], titleHTML: String, subtitle: String?, preview: String?, uri: String, subitemID: Int) {
+    public init(id: Int64, navSectionID: Int64, position: Int, imageRenditions: [ImageRendition], titleHTML: String, subtitle: String?, preview: String?, uri: String, subitemID: Int64) {
         self.id = id
         self.navSectionID = navSectionID
         self.position = position
